Renames clients, removes old ones
