+++ src/Registry.test.js
