+++ src/index.js
@@ -1,21 +1,17 @@
-import AppEngineClient from './AppEngineClient.js'
-import AppRegistryClient from './AppRegistryClient.js'
-import AppsClient from './AppsClient.js'
+import Apps from './Apps.js'
 import Client from './Client.js'
-import RegistryClient from './RegistryClient.js'
-import RouterClient from './RouterClient.js'
-import VBaseClient from './VBaseClient.js'
-import VTEXIDClient from './VTEXIDClient.js'
-import WorkspacesClient from './WorkspacesClient.js'
+import Registry from './Registry.js'
+import Router from './Router.js'
+import VBase from './VBase.js'
+import ID from './ID.js'
+import Workspaces from './Workspaces.js'
 
 module.exports = {
-  AppEngineClient,
-  AppRegistryClient,
-  AppsClient,
+  Apps,
   Client,
-  RegistryClient,
-  RouterClient,
-  VBaseClient,
-  VTEXIDClient,
-  WorkspacesClient,
+  Registry,
+  Router,
+  VBase,
+  ID,
+  Workspaces,
 }
